HubbleSee/requirements.txt
lxml
PyQt5
openpyxl
requests
pillow